+++ src/e2e/app/cli/repo-tests/src/rest_api/mod.rs
@@ -10,6 +10,7 @@
 mod test_accounts;
 mod test_auth;
 mod test_dataset;
+mod test_gql_query;
 mod test_odf_core;
 mod test_odf_query;
 mod test_openapi;
@@ -19,6 +20,7 @@ mod test_upload;
 pub use test_accounts::*;
 pub use test_auth::*;
 pub use test_dataset::*;
+pub use test_gql_query::*;
 pub use test_odf_core::*;
 pub use test_odf_query::*;
 pub use test_openapi::*;